+++ packages/std/src/addresses.rs
@@ -52,6 +52,11 @@ impl Addr {
     pub fn as_str(&self) -> &str {
         self.0.as_str()
     }
+
+    #[inline]
+    pub fn as_bytes(&self) -> &[u8] {
+        self.0.as_bytes()
+    }
 }
 
 impl fmt::Display for Addr {
@@ -285,6 +290,15 @@ mod tests {
     fn addr_as_str_works() {
         let addr = Addr::unchecked("literal-string");
         assert_eq!(addr.as_str(), "literal-string");
+    }
+
+    #[test]
+    fn addr_as_bytes_works() {
+        let addr = Addr::unchecked("literal-string");
+        assert_eq!(
+            addr.as_bytes(),
+            [108, 105, 116, 101, 114, 97, 108, 45, 115, 116, 114, 105, 110, 103]
+        );
     }
 
     #[test]